requirements.txt
requests
feedparser
trafilatura
beautifulsoup4
pyyaml
python-slugify